requirements.txt
langchain>=0.2
langchain-community>=0.2
faiss-cpu
sentence-transformers
pypdf
tiktoken
python-dotenv
openai>=1.40
streamlit